onors/requirements.txt
boto3==1.7.78
botocore==1.10.78
Django==2.1
django-cors-headers==2.4.0
django-storages==1.6.6
djangorestframework==3.8.2
docutils==0.14
jmespath==0.9.3
mysqlclient==1.3.13
Pillow==5.2.0
pkg-resources==0.0.0
python-dateutil==2.7.3
pytz==2018.5
s3transfer==0.1.13
six==1.11.0
